event_manager/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
defusedxml==0.7.1
Django==5.2.5
django-filter==25.1
django-taggit==6.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-yasg==1.21.11
idna==3.10
inflection==0.5.1
mysql-connector-python==9.4.0
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
pycparser==2.23
PyJWT==2.10.1
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
